Execute liteDB in Automate Tab

+++ Ginger/GingerCoreNET/Run/RunListenerLib/LiteDBRepository.cs
@@ -11,6 +11,7 @@ using Ginger.Run;
 using GingerCore;
 using GingerCore.Activities;
 using GingerCore.Environments;
+using LiteDB;
 
 namespace Amdocs.Ginger.CoreNET.Run.RunListenerLib
 {
@@ -28,6 +29,7 @@ namespace Amdocs.Ginger.CoreNET.Run.RunListenerLib
         {
             liteDbManager = new LiteDbManager(executionLoggerHelper.GetLoggerDirectory(WorkSpace.Instance.Solution.ExecutionLoggerConfigurationSetList.ExecutionLoggerConfigurationExecResultsFolder));
         }
+
         public override void SaveObjToReporsitory(object obj, string FileName = "", bool toAppend = false)
         {
             liteDbManager.WriteToLiteDb(FileName, new List<LiteDbReportBase>() { (LiteDbReportBase)obj });
@@ -37,11 +39,14 @@ namespace Amdocs.Ginger.CoreNET.Run.RunListenerLib
             string executionLogFolder = executionLoggerHelper.GetLoggerDirectory(WorkSpace.Instance.Solution.ExecutionLoggerConfigurationSetList.ExecutionLoggerConfigurationExecResultsFolder);
             LiteDbAction liteDbAction = new LiteDbAction();
             liteDbAction.SetReportData(GetActionReportData(action, context, executedFrom));
-            if (System.IO.Directory.Exists(executionLogFolder))
-            {
+            //if (System.IO.Directory.Exists(executionLogFolder))
+            //{
                 liteDbAction.Wait = action.Wait;
                 liteDbAction.TimeOut = action.Timeout;
-                
+                if (action.LiteDbId != null && executedFrom == eExecutedFrom.Automation)
+                {
+                    liteDbAction._id = action.LiteDbId;
+                }
                 // Save screenShots
                 int screenShotCountPerAction = 0;
                 for (var s = 0; s < action.ScreenShots.Count; s++)
@@ -72,12 +77,16 @@ namespace Amdocs.Ginger.CoreNET.Run.RunListenerLib
                 }
                 liteDbAction.ScreenShots = action.ScreenShots;
                 liteDbActionList.Add(liteDbAction);
-            }
-            else
-            {
-                Reporter.ToLog(eLogLevel.ERROR, "Failed to create ExecutionLogger JSON file for the Action :" + action.Description + " because directory not exists :" + executionLogFolder + action.ExecutionLogFolder);
-            }
+            //}
+            //else
+            //{
+            //    Reporter.ToLog(eLogLevel.ERROR, "Failed to create ExecutionLogger JSON file for the Action :" + action.Description + " because directory not exists :" + executionLogFolder + action.ExecutionLogFolder);
+            //}
             SaveObjToReporsitory(liteDbAction, liteDbManager.NameInDb<LiteDbAction>());
+            if (executedFrom == eExecutedFrom.Automation)
+            {
+                action.LiteDbId = liteDbAction._id;
+            }
             return liteDbAction;
         }
 
@@ -86,10 +95,18 @@ namespace Amdocs.Ginger.CoreNET.Run.RunListenerLib
             LiteDbActivity AR = new LiteDbActivity();
             AR.SetReportData(GetActivityReportData(activity,context, offlineMode));
             AR.ActivityGroupName = activity.ActivitiesGroupID;
+            if(activity.LiteDbId != null && ExecutionLoggerManager.RunSetReport.RunSetExecutionStatus == Execution.eRunStatus.Automated) // missing Executed from
+            {
+                AR._id = activity.LiteDbId;
+            }
             AR.ActionsColl.AddRange(liteDbActionList);
             liteDbActivityList.Add(AR);
             SaveObjToReporsitory(AR, liteDbManager.NameInDb<LiteDbActivity>());
             liteDbActionList.Clear();
+            if (ExecutionLoggerManager.RunSetReport.RunSetExecutionStatus == Execution.eRunStatus.Automated)
+            {
+                activity.LiteDbId = AR._id;
+            }
             return AR;
         }
         public override object SetReportActivityGroup(ActivitiesGroup activityGroup, BusinessFlow businessFlow, bool offlineMode)
@@ -106,6 +123,10 @@ namespace Amdocs.Ginger.CoreNET.Run.RunListenerLib
         {
             LiteDbBusinessFlow BFR = new LiteDbBusinessFlow();
             BFR.SetReportData(GetBFReportData(businessFlow, environment));
+            if(businessFlow.LiteDbId != null && executedFrom == eExecutedFrom.Automation)
+            {
+                BFR._id = businessFlow.LiteDbId;
+            }
             if (WorkSpace.Instance.Solution.ExecutionLoggerConfigurationSetList.ExecutionLoggerConfigurationIsEnabled)
             {
                 if (offlineMode)
@@ -141,6 +162,10 @@ namespace Amdocs.Ginger.CoreNET.Run.RunListenerLib
                     //this.BFCounter = 0;
                 }
             }
+            if (executedFrom == eExecutedFrom.Automation)
+            {
+                businessFlow.LiteDbId = BFR._id;
+            }
             return BFR;
         }
 
@@ -158,7 +183,7 @@ namespace Amdocs.Ginger.CoreNET.Run.RunListenerLib
             liteDbBFList.Clear();
         }
 
-        internal override void SetReportRunSet(RunSetReport runSetReport, string logFolder)
+        public override void SetReportRunSet(RunSetReport runSetReport, string logFolder)
         {
             LiteDbRunSet runSet = new LiteDbRunSet();
             base.SetReportRunSet(runSetReport, logFolder);
@@ -166,10 +191,34 @@ namespace Amdocs.Ginger.CoreNET.Run.RunListenerLib
             runSet.SetReportData(runSetReport);
             SaveObjToReporsitory(runSet, liteDbManager.NameInDb<LiteDbRunSet>());
             ExecutionLoggerManager.RunSetReport.liteDbRunnerList.Clear();
-            if(System.IO.Directory.Exists(runSetReport.LogFolder))
+            if(runSetReport.LogFolder != null && System.IO.Directory.Exists(runSetReport.LogFolder))
             {
                 System.IO.Directory.Delete(runSetReport.LogFolder,true);
             }
+        }
+        public override void RunSetUpdate(ObjectId runSetLiteDbId, ObjectId runnerLiteDbId, GingerRunner gingerRunner)
+        {
+            LiteDbRunner runner = new LiteDbRunner();
+            runner.BusinessFlowsColl.AddRange(liteDbBFList);
+            runner._id = runnerLiteDbId;
+            runner.Seq = 1;
+            runner.Name = "Automated Runner";
+            runner.ApplicationAgentsMappingList = gingerRunner.ApplicationAgents.Select(a => a.AgentName + "_:_" + a.AppName).ToList();
+            runner.Environment = gingerRunner.ProjEnvironment != null ? gingerRunner.ProjEnvironment.Name : string.Empty;
+            runner.GUID = gingerRunner.Guid;
+            if(gingerRunner.BusinessFlows.Count > 0)
+            {
+                runner.StartTimeStamp = gingerRunner.BusinessFlows[0].StartTimeStamp;
+                runner.EndTimeStamp = gingerRunner.BusinessFlows[0].EndTimeStamp;
+                runner.Elapsed = gingerRunner.BusinessFlows[0].Elapsed;
+            }
+            SaveObjToReporsitory(runner, liteDbManager.NameInDb<LiteDbRunner>());
+            liteDbBFList.Clear();
+            LiteDbRunSet runSet = new LiteDbRunSet();
+            runSet._id = runSetLiteDbId;
+            base.SetReportRunSet(ExecutionLoggerManager.RunSetReport, "");
+            runSet.SetReportData(ExecutionLoggerManager.RunSetReport);
+            SaveObjToReporsitory(runSet, liteDbManager.NameInDb<LiteDbRunSet>());
         }
     }
 }
